requirements.txt
APScheduler==3.6.3
codacy-coverage==1.3.11
configparser==4.0.2
coverage==4.5.1
Flask==1.1.2
Flask-Compress==1.5.0
Flask-Migrate==2.5.3
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.8
Flask-Session==0.3.2
Flask-SQLAlchemy==2.4.4
Flask-SSLify==0.1.5
Flask-Sockets==0.2.1
GitHub-Flask==3.2.0
gunicorn==20.0.4
mock==3.0.5
psycopg2==2.8.6
pycodestyle==2.6.0
PyGithub==1.51
pytest==5.4.3
pytest-pythonpath==0.7.3
pytest-cov==2.10.0
requests==2.24.0
SQLAlchemy-Utils==0.36.8
cryptography==3.2.1
